refactor(stella-cli): split candidate_ws tests into candidate_ws/tests.rs
candidate_ws.rs crossed the 1500-line file-size ratchet (1533 lines) once
#2067's baseline-pinning tests landed, and the ratchet accepts no new
baseline entries. Move the inline #[cfg(test)] module to a sibling file,
the same shape as agent.rs / agent/tests.rs. Pure move: no test added,
renamed, or dropped.

Refs #2067
